src/europa/poller/requirements.txt
requests==2.18.4
Adafruit-DHT==1.3.2
w1thermsensor==1.0.5
RPi.GPIO==0.6.3
pyhs100==0.3.0
